start extracting things into modules
The borrow_check module is too big for its own good
